Refactor swagger support and introduce much-needed tests
